noop component rename
